MorphoStream.java: the "debug" parameter wasn't actually getting set in the constructor, fixed that.
HMM.java: Fixed some typos, added missing code. Added *alot* of missing code to train(), but didn't finish it. Apparently that's not used currently. ^^;
Added a new exception at the beginning of train() to remind people that it's only in a half-finished state. (The if(true) statement is so that it will compile.)