Added rm call to TransferEnEoTest and TransferEoEnTest to remove compiled bin file from tmp dir, as leaving it there was causing side effects with other tests sometimes. Added another string to the StringTable, will simplify it in next commit.
Refactored Transfer compile (which is also used by Interchunk and Postchunk) to split the actual compilation off into separate classes, and make it work with the pipeline better. Right now there are two options, InternalTransferCompile and ExternalTransferCompile, which are tried in that order.

InternalTransferCompile uses ToolProvider.getSystemJavaCompiler(), while ExternalTransferCompile uses exec.
The former works when the runtime environment we are in is a JDK, but does not work when running in a JRE.
If the internal fails, fallback to the external, which may still be accessible if we're running in a JRE, but the user has a JDK on the system path.
If external fails, well not much we can do about that except throw an exception. ;)