Added support for negative tests: make sure that a certain morphosyntactic form does NOT get generated. This is relevant for words with defective paradigms.
Example (from SMJ, Julev Saami):

Config:
  hfst:
    Gen: ../../../src/generator.gt.hfst
    Morph: ../../../src/analyser.gt.hfst
  xerox:
    Gen: ../../../src/generator.gt.xfst
    Morph: ../../../src/analyser.gt.xfst
    App: lookup

Tests:

  Adjective - antisosiálla: #  ÁLLA
    antisosiálla+A+Attr: antisosiála
    antisosiálla+A+Sg+Nom: antisosiálla
    antisosiálla+A+Sg+Gen: antisosiála
    antisosiálla+A+Comp+Sg+Nom: []
    antisosiálla+A+Superl+Sg+Nom: []


I don't have the infrastructure right now to test that the code is working as it should in Apertium, but the code is working fine in the Giellatekno infrastructure.

Please test and give feedback in IRC #hfst (or #apertium if I happen to lurk there).